Separated CtS, StC, SinEval from the bootstrapping parameters
